Fall back to storage upload if overwriting non mp4
In cases where the filename was already present in storage, and it is
not in the encoding format, allow overwrite of the original file directly
to GCS.

Otherwise, we'd encode and save the file as mp4, and the
original would not be overwritten. This would be a problem when users
intend to overwrite content with fresh content using the same file name.